perf: don't store keys array on node instance
This is another instance of reducing the memory footprint of the AST
node base class. The keys array is a cached array of which properties on
the node correspond to branches of the syntax tree based on the original
esTree nodes. The keys property of the NodeBase class is a reference to
this cached array. While this is a nice convenience, it is not necessary
and bloats the size of each AST node instance.

Instead of storing a reference to this array on each node instance, we
can instead use the node's type field to look up the keys array when we
need it since it is already stored in an map from type=>keys.