perf: avoid storing AstContext on base AST Node
Another instance of reducing AST node bloat. The context property of the
NodeBase class increases the size of every AST node. This commit moves
the context property from the NodeBase class to the ChildScope class,
and since every node has a scope property we can access the context
through an additional property access `this.scope.context`.

Generally speaking, we can expect that number of scope objects should be
much smaller than the total number of ast node objects. So while the
scope objects grow in size a little bit, the total memory usage should
be reduced as a result of this change.